Combine Tracker and Library into one Reforms surface
Both pages were lists of reform-shaped objects with the same endgame
(open a report or a draft) — the difference is provenance, which is
now a badge, not a page:

- Reforms page: one list with filter chips (All · In Congress ·
  Yours) and search across both kinds. Bill rows carry jurisdiction +
  status badges and expand to summary, provision table, "View full
  impact report", and "Open as draft reform" (fork). Your rows carry
  a "Yours" badge and expand to the amend-in-place editor (rename,
  values, save, duplicate, delete, add-in-Build) plus the same
  report button.
- Sidebar tightens to Ask · Build · Reforms; Recent entries land on
  the Yours filter. Ask's tiles deep-link to the filters.
- /tracker and /library redirect to /reforms preserving queries
  (?bill= deep links still open the bill expanded); saving a draft
  lands on Yours. Old pages and tests deleted; 12 merged tests cover
  filters, deep links, forking, amending, and search across kinds.
