+++ app/src/CalculatorRouter.tsx
@@ -2,15 +2,21 @@
  * Router for the Calculator app (app.policyengine.org)
  * Contains only the interactive calculator functionality
  */
-import { createBrowserRouter, Navigate, Outlet, RouterProvider, useParams } from 'react-router-dom';
+import {
+  createBrowserRouter,
+  Navigate,
+  Outlet,
+  RouterProvider,
+  useLocation,
+  useParams,
+} from 'react-router-dom';
 import PathwayLayout from './components/PathwayLayout';
 import StandardLayout from './components/StandardLayout';
 import { isFlagshipShellEnabled } from './libs/featureFlags';
 import AskPage from './pages/flagship/Ask.page';
 import BuildPage from './pages/flagship/Build.page';
-import LibraryPage from './pages/flagship/Library.page';
+import ReformsPage from './pages/flagship/Reforms.page';
 import FlagshipReportPage from './pages/flagship/Report.page';
-import TrackerPage from './pages/flagship/Tracker.page';
 import NotFoundPage from './pages/NotFound.page';
 import PoliciesPage from './pages/Policies.page';
 import PopulationsPage from './pages/Populations.page';
@@ -26,6 +32,12 @@ import { CountryGuard } from './routing/guards/CountryGuard';
 import { MetadataGuard } from './routing/guards/MetadataGuard';
 import { MetadataLazyLoader } from './routing/guards/MetadataLazyLoader';
 import { RedirectToCountry } from './routing/RedirectToCountry';
+
+/** Tracker and Library merged into Reforms; preserve deep-link queries. */
+function LegacyReformsRedirect() {
+  const location = useLocation();
+  return <Navigate to={`../reforms${location.search}`} replace />;
+}
 
 /** Bridges react-router useParams to ModifyReportPage's prop interface. */
 function ModifyReportPageRoute() {
@@ -126,9 +138,10 @@ const router = createBrowserRouter(
                   ? [
                       { path: 'home', element: <Navigate to="../ask" replace /> },
                       { path: 'ask', element: <AskPage /> },
-                      { path: 'tracker', element: <TrackerPage /> },
                       { path: 'build', element: <BuildPage /> },
-                      { path: 'library', element: <LibraryPage /> },
+                      { path: 'reforms', element: <ReformsPage /> },
+                      { path: 'tracker', element: <LegacyReformsRedirect /> },
+                      { path: 'library', element: <LegacyReformsRedirect /> },
                       { path: 'report/:userReportId', element: <FlagshipReportPage /> },
                     ]
                   : []),
