+++ calculator-app/src/app/[countryId]/(calculator)/library/page.tsx
@@ -1,12 +1,20 @@
 "use client";
 
-import LibraryPage from "@/pages/flagship/Library.page";
-import FlagshipGate from "../FlagshipGate";
+import { use, useEffect } from "react";
+import { useRouter } from "next/navigation";
 
-export default function LibraryRoute() {
-  return (
-    <FlagshipGate>
-      <LibraryPage />
-    </FlagshipGate>
-  );
+/** Library merged into Reforms; land on the Yours filter. */
+export default function LibraryRoute({
+  params,
+}: {
+  params: Promise<{ countryId: string }>;
+}) {
+  const { countryId } = use(params);
+  const router = useRouter();
+
+  useEffect(() => {
+    router.replace(`/${countryId}/reforms?filter=yours`);
+  }, [router, countryId]);
+
+  return null;
 }
