+++ calculator-app/src/app/[countryId]/(calculator)/reforms/page.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+import ReformsPage from "@/pages/flagship/Reforms.page";
+import FlagshipGate from "../FlagshipGate";
+
+export default function ReformsRoute() {
+  return (
+    <FlagshipGate>
+      <ReformsPage />
+    </FlagshipGate>
+  );
+}
